+++ packages/tolk/lib/runtime/nv/tolk_nv.ml
@@ -1948,6 +1948,7 @@ module Allocator = struct
       free;
       copyin = copyin state;
       copyout = copyout state;
+      as_buffer = None;
       addr = Hcq.Buffer.va;
       offset = Some offset;
       transfer = Some (transfer state);